Firefox replaces Google Chrome.
dopamine-2020 started with my upgrading to i3 4.18 on xubuntu 20.04
LTS. Google Chrome has some issues at startup, and is third-party. Firefox
is well behaved and ships with the distro. Firefox has the ten aliases,
Google just the one, to remind me which browser to work with. Small
steps, but the apps array ought to be loaded so that updates don't need
editing. Stay tuned.