PoC demo of inserting a file into another file.
The demo has no effect on on the running system
at the moment. Small steps.

(1)
i3scripts/i3-focus-app-by-alias
is available in
myscripts/my-focus-app-by-alias
with the apps table entries stripped out and replaced with
INSERTAPPSHERE,
The apps table entries are stored in the file my-apps.

(2)
Typing
'make apps'
replaces the
INSERTAPPSHERE
with the apps table

(3)
diff \
i3scripts/i3-focus-app-by-alias \
myscripts/my-focus-app-by-alias
confirms that the insertion restores the original.