Add "mousewap", remove "floating" on window event.
The "floating" function remains but is not called until I find the
bug that fails to render (mumble).

The "mousewarp" is a silly function to put the mouse cursor inside themm
focused window even when using just the keyboard to change focus. I have
tried this before, it can be overzealous.